Reprojecting all images and using crsTransform instead of scale
Remove the filterBounds calls on the CONUS assets.

Removed the ee_compute_etdata function since it seemed easier to just have this one call in the main loop.

Building the tile ee.Feature one time in the main loop instead of a second time in set_ee_etdata_img, and passing the tile feature into the functions instead of the tile ID.

Using the default Landsat tile transform instead of scale.